@@ -28,45 +29,50 @@
     "dep-graph": "nx dep-graph",
     "help": "nx help",
     "workspace-generator": "nx workspace-generator",
+    "demo.android": "nx run nativescript-demo-ng:android",
+    "demo.ios": "nx run nativescript-demo-ng:ios",
+    "demo.clean": "nx run nativescript-demo-ng:clean",
     "changelog": "conventional-changelog -p angular -i CHANGELOG.md -s"
   },
   "dependencies": {
-    "@angular/animations": "12.2.5",
-    "@angular/common": "12.2.5",
-    "@angular/compiler": "12.2.5",
-    "@angular/core": "12.2.5",
-    "@angular/forms": "12.2.5",
-    "@angular/platform-browser": "12.2.5",
-    "@angular/platform-browser-dynamic": "12.2.5",
-    "@angular/router": "12.2.5",
+    "@angular/animations": "13.1.1",
+    "@angular/common": "13.1.1",
+    "@angular/compiler": "13.1.1",
+    "@angular/core": "13.1.1",
+    "@angular/forms": "13.1.1",
+    "@angular/platform-browser": "13.1.1",
+    "@angular/platform-browser-dynamic": "13.1.1",
+    "@angular/router": "13.1.1",
     "@nativescript/core": "~8.1.0",
     "@nativescript/theme": "~3.0.0",
+    "@ngx-translate/core": "~14.0.0",
     "nativescript-ngx-fonticon": "~7.0.0",
-    "rxjs": "^7.3.0",
-    "zone.js": "0.11.4"
+    "rxjs": "^7.4.0",
+    "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-eslint/eslint-plugin": "12.3.1",
-    "@angular-eslint/eslint-plugin-template": "12.3.1",
-    "@angular-eslint/template-parser": "12.3.1",
-    "@angular/compiler-cli": "12.2.5",
+    "@angular-devkit/build-angular": "~13.1.0",
+    "@angular-eslint/eslint-plugin": "13.0.1",
+    "@angular-eslint/eslint-plugin-template": "13.0.1",
+    "@angular-eslint/template-parser": "13.0.1",
+    "@angular/compiler-cli": "13.1.1",
     "@nativescript/types": "~8.1.0",
     "@nativescript/unit-test-runner": "^2.0.5",
-    "@nativescript/webpack": "5.0.0",
-    "@ngtools/webpack": "12.2.5",
-    "@nrwl/angular": "12.9.0",
-    "@nrwl/cli": "12.9.0",
-    "@nrwl/eslint-plugin-nx": "12.9.0",
-    "@nrwl/jest": "12.9.0",
-    "@nrwl/node": "12.9.0",
-    "@nrwl/tao": "12.9.0",
-    "@nrwl/workspace": "12.9.0",
-    "@nstudio/angular": "12.7.0",
-    "@nstudio/nativescript": "12.7.0",
-    "@nstudio/nativescript-angular": "12.7.0",
-    "@nstudio/xplat": "12.7.0",
+    "@nativescript/webpack": "5.0.2-alpha.0",
+    "@ngtools/webpack": "next",
+    "@nrwl/angular": "13.3.12",
+    "@nrwl/cli": "13.3.12",
+    "@nrwl/eslint-plugin-nx": "13.3.12",
+    "@nrwl/jest": "13.3.12",
+    "@nrwl/node": "13.3.12",
+    "@nrwl/tao": "13.3.12",
+    "@nrwl/workspace": "13.3.12",
+    "@nstudio/angular": "13.0.1",
+    "@nstudio/nativescript": "13.0.1",
+    "@nstudio/nativescript-angular": "13.0.1",
+    "@nstudio/xplat": "13.0.1",
     "@types/jasmine": "^3.7.2",
-    "@types/jest": "~26.0.8",
+    "@types/jest": "27.0.2",
     "@types/node": "14.14.33",
     "@types/sinon": "^10.0.2",
     "@typescript-eslint/eslint-plugin": "4.19.0",
@@ -85,14 +91,18 @@
     "karma-sinon": "^1.0.5",
     "karma-webpack": "5.0.0",
     "lint-staged": "^11.0.0",
-    "ng-packagr": "12.2.1",
+    "ng-packagr": "13.0.8",
+    "postcss": "^8.3.9",
+    "postcss-import": "^14.0.2",
+    "postcss-preset-env": "^6.7.0",
+    "postcss-url": "^10.1.1",
     "prettier": "2.3.2",
     "sass": "^1.32.0",
     "sinon": "^11.1.1",
-    "ts-jest": "27.0.3",
+    "ts-jest": "27.0.5",
     "ts-node": "9.1.1",
     "tslib": "^2.1.0",
-    "typescript": "4.3.5"
+    "typescript": "4.4.4"
   },
   "xplat": {
     "prefix": "nativescript",
