@@ -7,10 +7,28 @@ import { ItemsComponent } from './item/items.component';
 import { ItemDetailComponent } from './item/item-detail.component';
 import { ModalComponent } from './modal/modal.component';
 
+/**
+ * To test tab named outlets, can uncomment imports and declarations
+ */
+// import { HomeComponent } from './home/home.component';
+// import { NativeScriptMaterialBottomNavigationModule } from '@nativescript-community/ui-material-bottom-navigation/angular';
+
 @NgModule({
   bootstrap: [AppComponent],
-  imports: [NativeScriptModule, NativeScriptHttpClientModule, AppRoutingModule, NativeDialogModule],
-  declarations: [AppComponent, ItemsComponent, ItemDetailComponent, ModalComponent],
+  imports: [
+    NativeScriptModule,
+    NativeScriptHttpClientModule,
+    AppRoutingModule,
+    NativeDialogModule,
+    // NativeScriptMaterialBottomNavigationModule
+  ],
+  declarations: [
+    AppComponent,
+    ItemsComponent,
+    ItemDetailComponent,
+    ModalComponent,
+    // HomeComponent
+  ],
   providers: [],
   schemas: [NO_ERRORS_SCHEMA],
 })
